refactor(#147): auth에서 user로 이동하면서 옮겨진 파일들
