--- exbackend/src/domain/auth/dto/index.ts
@@ -1,2 +1,0 @@
-export { UserEmailDto } from './user-email.dto';
-export { ManageMemberDto } from './manage-member.dto';
