--- exbackend/src/domain/user/controllers/user.controller.ts
@@ -1,11 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-import { UserService } from '../services/user.service';
-
-@ApiTags('Users')
-@Controller('users')
-export class UserController {
-  constructor(private readonly userService: UserService) {}
-
-  // TODO: API 엔드포인트 구현 예정
-}
