+++ exbackend/src/domain/user/index.ts
@@ -1,0 +1,2 @@
+export * from './decorators/current-user.decorator';
+export * from './entities/user.entity';
