+++ exbackend/src/domain/user/user-activity/last-channel-info.dto.ts
